Preventing crash when updating table view on non-active controller.
Without this, `Row` would crash during `didEndDisplayCell` handling if table view is not on active controller (for example if there's modal controller on top of table view controller).

The solution here is to simply ignore the call if `didEndDisplayCell` block is not assigned on `Row`. This preserves existing functionality for cases where block is used but prevents the crash otherwise.